backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.66
pandas>=2.2.0
numpy>=1.26.0
openai==2.6.0
python-dotenv==1.1.1
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot==20.8
schedule==1.2.0
